+++ buildcontext/local.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/earthly/earthly/domain"
 	"github.com/earthly/earthly/gitutil"
-	"github.com/earthly/earthly/llbutil"
-	"github.com/earthly/earthly/llbutil/pllb"
+	"github.com/earthly/earthly/util/llbutil"
+	"github.com/earthly/earthly/util/llbutil/pllb"
 	"github.com/earthly/earthly/util/syncutil/synccache"
 	"github.com/moby/buildkit/client/llb"
 	"github.com/pkg/errors"
