+++ earthfile2llb/runmount.go
@@ -5,9 +5,9 @@ import (
 	"strings"
 
 	"github.com/earthly/earthly/domain"
-	"github.com/earthly/earthly/llbutil"
-	"github.com/earthly/earthly/llbutil/pllb"
 	"github.com/earthly/earthly/states/dedup"
+	"github.com/earthly/earthly/util/llbutil"
+	"github.com/earthly/earthly/util/llbutil/pllb"
 	"github.com/moby/buildkit/client/llb"
 	"github.com/pkg/errors"
 )
