+++ earthfile2llb/withdockerrun.go
@@ -14,9 +14,9 @@ import (
 	"github.com/containerd/containerd/platforms"
 	"github.com/earthly/earthly/dockertar"
 	"github.com/earthly/earthly/domain"
-	"github.com/earthly/earthly/llbutil"
-	"github.com/earthly/earthly/llbutil/pllb"
 	"github.com/earthly/earthly/states"
+	"github.com/earthly/earthly/util/llbutil"
+	"github.com/earthly/earthly/util/llbutil/pllb"
 	"github.com/moby/buildkit/client/llb"
 	gwclient "github.com/moby/buildkit/frontend/gateway/client"
 	specs "github.com/opencontainers/image-spec/specs-go/v1"
