+++ util/llbutil/copyop.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/earthly/earthly/llbutil/pllb"
+	"github.com/earthly/earthly/util/llbutil/pllb"
 	"github.com/moby/buildkit/client/llb"
 	"github.com/pkg/errors"
 )
