+++ util/llbutil/fakedep.go
@@ -1,7 +1,7 @@
 package llbutil
 
 import (
-	"github.com/earthly/earthly/llbutil/pllb"
+	"github.com/earthly/earthly/util/llbutil/pllb"
 	"github.com/moby/buildkit/client/llb"
 )
